requirements.txt
django==4.2
django-allauth==0.52.0
django-crispy-forms==1.12
django-environ==0.11.2
django-chunk-upload==2.0.0
celery==4.2.0
gevent==23.7.0
redis==4.5.4
djangorestframework==3.14.0